+++ packages/trackback-agent/src/agent/index.ts
@@ -1,6 +1,7 @@
 import { ITrackBackAgent, IConnect, ITrackbackAccount } from '../types';
 import { IProcedure, Procedure } from './procedure';
 import { createAccount } from './account';
+import { DecentralisedFileStoreConnector } from './connection';
 
 export * from './connection'
 export * from './account'
@@ -11,8 +12,8 @@ export class TrackBackAgent implements ITrackBackAgent {
 
   procedure: IProcedure;
 
-  constructor(connection: IConnect | null | undefined) {
-    this.procedure = new Procedure(connection);
+  constructor(connection?: IConnect, fileConnector?: DecentralisedFileStoreConnector) {
+    this.procedure = new Procedure(connection, fileConnector);
   }