+++ packages/trackback-did/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.1-alpha.9](https://github.com/trackback-blockchain/trackback-did/compare/@trackback/did@0.0.1-alpha.8...@trackback/did@0.0.1-alpha.9) (2021-11-18)
+
+**Note:** Version bump only for package @trackback/did
+
+
+
+
+
+## [0.0.1-alpha.8](https://github.com/trackback-blockchain/trackback-did/compare/@trackback/did@0.0.1-alpha.7...@trackback/did@0.0.1-alpha.8) (2021-11-18)
+
+**Note:** Version bump only for package @trackback/did
+
+
+
+
+
+## [0.0.1-alpha.7](https://github.com/trackback-blockchain/trackback-did/compare/@trackback/did@0.0.1-alpha.6...@trackback/did@0.0.1-alpha.7) (2021-11-18)
+
+**Note:** Version bump only for package @trackback/did
+
+
+
+
+
+## [0.0.1-alpha.6](https://github.com/trackback-blockchain/trackback-did/compare/@trackback/did@0.0.1-alpha.5...@trackback/did@0.0.1-alpha.6) (2021-11-15)
+
+**Note:** Version bump only for package @trackback/did
+
+
+
+
+
+## [0.0.1-alpha.5](https://github.com/trackback-blockchain/trackback-did/compare/@trackback/did@0.0.1-alpha.4...@trackback/did@0.0.1-alpha.5) (2021-10-28)
+
+**Note:** Version bump only for package @trackback/did
+
+
+
+
+
 ## [0.0.1-alpha.4](https://github.com/trackback-blockchain/trackback-did/compare/@trackback/did@0.0.1-alpha.3...@trackback/did@0.0.1-alpha.4) (2021-10-28)
 
 **Note:** Version bump only for package @trackback/did